LRDOCS-522 Rename to PropertiesDocBuilder and create new dir for this
