+++ portal-impl/src/com/liferay/portal/tools/propertiesdoc/PropertiesSection.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.tools;
+package com.liferay.portal.tools.propertiesdoc;
 
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.tools.PropertiesConverter.PropertyComment;
+import com.liferay.portal.tools.propertiesdoc.PropertiesDocBuilder.PropertyComment;
 
 import java.util.List;
