+++ portal-impl/src/com/liferay/portal/tools/propertiesdoc/dependencies/properties_html.ftl
